Changed block type from char to unsigned char

+++ ibrdtn/ibrdtn/ibrdtn/data/Bundle.h
@@ -78,8 +78,8 @@ namespace dtn
 
 				const std::set<dtn::data::EID> getEIDs() const;
 
-				Block& get(int index);
-				const Block& get(int index) const;
+				Block& get(size_t index);
+				const Block& get(size_t index) const;
 				template<class T> T& get();
 				template<class T> const T& get() const;
 
@@ -107,8 +107,8 @@ namespace dtn
 
 			const block_list& getBlocks() const;
 
-			dtn::data::Block& getBlock(int index);
-			const dtn::data::Block& getBlock(int index) const;
+			dtn::data::Block& getBlock(size_t index);
+			const dtn::data::Block& getBlock(size_t index) const;
 
 			template<class T>
 			T& getBlock();
